refactor(cli): reorganize module structure for SRP and extensibility
Split oversized modules into focused submodules:
- di/ → wiring/ (repositories, use_cases, app_state, dns/{cache,pool,resolver})
- server/dns.rs → server/dns/{mod,udp,tcp,pktinfo}
- main.rs: 275→129 lines; extracted args, jobs, app_state

Added DnsProtocolServer trait as extension point for future protocols (DoT, DoH).
All files kept under 200 lines. Zero logic changes.
