+++ crates/cli/src/bootstrap/mod.rs
@@ -1,7 +1,9 @@
 pub mod config;
 pub mod database;
+pub mod jobs;
 pub mod logging;
 
 pub use config::load_config;
 pub use database::init_database;
+pub use jobs::build_job_runner;
 pub use logging::init_logging;
