+++ crates/cli/src/main.rs
@@ -4,54 +4,18 @@ use mimalloc::MiMalloc;
 static GLOBAL: MiMalloc = MiMalloc;
 
 use clap::Parser;
-use ferrous_dns_api::{
-    AppState, BlockingUseCases, ClientUseCases, DnsUseCases, GroupUseCases, QueryUseCases,
-    SafeSearchUseCases, ScheduleUseCases, ServiceUseCases,
-};
-use ferrous_dns_application::use_cases::{
-    CreateLocalRecordUseCase, DeleteLocalRecordUseCase, UpdateLocalRecordUseCase,
-};
 use ferrous_dns_domain::CliOverrides;
 use ferrous_dns_infrastructure::dns::server::DnsServerHandler;
-use ferrous_dns_infrastructure::dns::UpstreamHealthAdapter;
-use ferrous_dns_infrastructure::repositories::{SqliteConfigRepository, TomlConfigFilePersistence};
-use ferrous_dns_jobs::{
-    BlocklistSyncJob, CacheMaintenanceJob, ClientSyncJob, JobRunner, QueryLogRetentionJob,
-    RetentionJob, ScheduleEvaluatorJob, WalCheckpointJob,
-};
 use std::net::SocketAddr;
 use std::sync::atomic::{AtomicUsize, Ordering};
 use std::sync::Arc;
 use tokio::sync::RwLock;
 use tracing::{error, info};
 
+mod args;
 mod bootstrap;
-mod di;
 mod server;
-
-#[derive(Parser)]
-#[command(name = "ferrous-dns")]
-#[command(version = "0.1.0")]
-#[command(about = "Ferrous DNS - High-performance DNS server with ad-blocking")]
-struct Cli {
-    #[arg(short = 'c', long, value_name = "FILE")]
-    config: Option<String>,
-
-    #[arg(short = 'd', long)]
-    dns_port: Option<u16>,
-
-    #[arg(short = 'w', long)]
-    web_port: Option<u16>,
-
-    #[arg(short = 'b', long)]
-    bind: Option<String>,
-
-    #[arg(long)]
-    database: Option<String>,
-
-    #[arg(long)]
-    log_level: Option<String>,
-}
+mod wiring;
 
 fn main() -> anyhow::Result<()> {
     let core_ids = core_affinity::get_core_ids().unwrap_or_default();
@@ -81,7 +45,7 @@ fn main() -> anyhow::Result<()> {
 }
 
 async fn async_main() -> anyhow::Result<()> {
-    let cli = Cli::parse();
+    let cli = args::Cli::parse();
 
     let cli_overrides = CliOverrides {
         dns_port: cli.dns_port,
@@ -108,40 +72,21 @@ async fn async_main() -> anyhow::Result<()> {
     let config_repo_pool = wal_pool.clone();
 
     let repos =
-        di::Repositories::new(write_pool, query_log_pool, read_pool, &config.database).await?;
-    let dns_services = di::DnsServices::new(&config, &repos).await?;
-    let use_cases = di::UseCases::new(
+        wiring::Repositories::new(write_pool, query_log_pool, read_pool, &config.database).await?;
+    let dns_services = wiring::DnsServices::new(&config, &repos).await?;
+    let use_cases = wiring::UseCases::new(
         &repos,
         dns_services.pool_manager.clone(),
         config.dns.local_dns_server.clone(),
     );
 
-    let mut runner = JobRunner::new()
-        .with_client_sync(ClientSyncJob::new(
-            use_cases.sync_arp.clone(),
-            use_cases.sync_hostnames.clone(),
-        ))
-        .with_retention(RetentionJob::new(use_cases.cleanup_clients.clone(), 30))
-        .with_query_log_retention(QueryLogRetentionJob::new(
-            use_cases.cleanup_query_logs.clone(),
-            config.database.queries_log_stored,
-        ))
-        .with_blocklist_sync(BlocklistSyncJob::new(repos.block_filter_engine.clone()))
-        .with_wal_checkpoint(WalCheckpointJob::new(
-            wal_pool,
-            config.database.wal_checkpoint_interval_secs,
-        ))
-        .with_schedule_evaluator(ScheduleEvaluatorJob::new(
-            repos.schedule_profile.clone(),
-            repos.schedule_state.clone(),
-        ));
-
-    if let Some(maintenance) = dns_services.cache_maintenance {
-        runner = runner.with_cache_maintenance(
-            CacheMaintenanceJob::new(maintenance)
-                .with_intervals(60, config.dns.cache_compaction_interval),
-        );
-    }
+    let runner = bootstrap::build_job_runner(
+        &use_cases,
+        &repos,
+        &config,
+        wal_pool,
+        dns_services.cache_maintenance.clone(),
+    );
 
     runner.start().await;
 
@@ -151,104 +96,14 @@ async fn async_main() -> anyhow::Result<()> {
     }
     info!("Subnet matcher cache loaded");
 
-    let config_repo: Arc<dyn ferrous_dns_application::ports::ConfigRepository> =
-        Arc::new(SqliteConfigRepository::new(config_repo_pool));
-
-    let app_state = AppState {
-        query: QueryUseCases {
-            get_stats: use_cases.get_stats,
-            get_queries: use_cases.get_queries,
-            get_timeline: use_cases.get_timeline,
-            get_query_rate: use_cases.get_query_rate,
-            get_cache_stats: use_cases.get_cache_stats,
-            get_top_blocked_domains: use_cases.get_top_blocked_domains,
-            get_top_clients: use_cases.get_top_clients,
-        },
-        dns: DnsUseCases {
-            cache: dns_services.cache.clone()
-                as Arc<dyn ferrous_dns_application::ports::DnsCachePort>,
-            create_local_record: Arc::new(CreateLocalRecordUseCase::new(
-                config_arc.clone(),
-                config_repo.clone(),
-            )),
-            update_local_record: Arc::new(UpdateLocalRecordUseCase::new(
-                config_arc.clone(),
-                config_repo.clone(),
-            )),
-            delete_local_record: Arc::new(DeleteLocalRecordUseCase::new(
-                config_arc.clone(),
-                config_repo,
-            )),
-            upstream_health: Arc::new(UpstreamHealthAdapter::new(
-                dns_services.pool_manager.clone(),
-                dns_services.health_checker.clone(),
-            )),
-        },
-        groups: GroupUseCases {
-            get_groups: use_cases.get_groups,
-            create_group: use_cases.create_group,
-            update_group: use_cases.update_group,
-            delete_group: use_cases.delete_group,
-            assign_client_group: use_cases.assign_client_group,
-        },
-        clients: ClientUseCases {
-            get_clients: use_cases.get_clients,
-            create_manual_client: use_cases.create_manual_client,
-            update_client: use_cases.update_client,
-            delete_client: use_cases.delete_client,
-            get_client_subnets: use_cases.get_client_subnets,
-            create_client_subnet: use_cases.create_client_subnet,
-            delete_client_subnet: use_cases.delete_client_subnet,
-            subnet_matcher: use_cases.subnet_matcher.clone(),
-        },
-        blocking: BlockingUseCases {
-            get_blocklist: use_cases.get_blocklist,
-            get_blocklist_sources: use_cases.get_blocklist_sources,
-            create_blocklist_source: use_cases.create_blocklist_source,
-            update_blocklist_source: use_cases.update_blocklist_source,
-            delete_blocklist_source: use_cases.delete_blocklist_source,
-            get_whitelist: use_cases.get_whitelist,
-            get_whitelist_sources: use_cases.get_whitelist_sources,
-            create_whitelist_source: use_cases.create_whitelist_source,
-            update_whitelist_source: use_cases.update_whitelist_source,
-            delete_whitelist_source: use_cases.delete_whitelist_source,
-            get_managed_domains: use_cases.get_managed_domains,
-            create_managed_domain: use_cases.create_managed_domain,
-            update_managed_domain: use_cases.update_managed_domain,
-            delete_managed_domain: use_cases.delete_managed_domain,
-            get_regex_filters: use_cases.get_regex_filters,
-            create_regex_filter: use_cases.create_regex_filter,
-            update_regex_filter: use_cases.update_regex_filter,
-            delete_regex_filter: use_cases.delete_regex_filter,
-            get_block_filter_stats: use_cases.get_block_filter_stats,
-        },
-        services: ServiceUseCases {
-            get_service_catalog: use_cases.get_service_catalog,
-            get_blocked_services: use_cases.get_blocked_services,
-            block_service: use_cases.block_service,
-            unblock_service: use_cases.unblock_service,
-            create_custom_service: use_cases.create_custom_service,
-            get_custom_services: use_cases.get_custom_services,
-            update_custom_service: use_cases.update_custom_service,
-            delete_custom_service: use_cases.delete_custom_service,
-        },
-        safe_search: SafeSearchUseCases {
-            get_configs: use_cases.get_safe_search_configs,
-            toggle: use_cases.toggle_safe_search,
-            delete_configs: use_cases.delete_safe_search_configs,
-        },
-        schedule: ScheduleUseCases {
-            get_profiles: use_cases.get_schedule_profiles,
-            create_profile: use_cases.create_schedule_profile,
-            update_profile: use_cases.update_schedule_profile,
-            delete_profile: use_cases.delete_schedule_profile,
-            manage_slots: use_cases.manage_time_slots,
-            assign_profile: use_cases.assign_schedule_profile,
-        },
-        config: config_arc.clone(),
-        config_file_persistence: Arc::new(TomlConfigFilePersistence),
-        api_key: config.server.api_key.as_deref().map(Arc::from),
-    };
+    let api_key = config.server.api_key.as_deref().map(Arc::from);
+    let app_state = wiring::build_app_state(
+        use_cases,
+        &dns_services,
+        config_arc,
+        config_repo_pool,
+        api_key,
+    );
 
     let dns_addr = format!("{}:{}", config.server.bind_address, config.server.dns_port);
     let dns_handler = DnsServerHandler::new(dns_services.handler_use_case);
