+++ crates/cli/src/wiring/mod.rs
@@ -1,7 +1,9 @@
+pub mod app_state;
 pub mod dns;
 pub mod repositories;
 pub mod use_cases;
 
+pub use app_state::build_app_state;
 pub use dns::DnsServices;
 pub use repositories::Repositories;
 pub use use_cases::UseCases;
